+ Split Helper Functions to New File
